moved data_loader to loading package and added documentation to data classes in loading package
