--- geoflow-core/src/main/kotlin/data_loader/AnalyzeInfo.kt
@@ -1,9 +1,0 @@
-package data_loader
-
-data class AnalyzeInfo(
-    val stOid: Long,
-    val tableName: String,
-    val delimiter: Char,
-    val qualified: Boolean,
-    val subTable: String?,
-)
